learning rate is float

+++ utils/arguments.py
@@ -14,7 +14,7 @@ def train_args():
     parser.add_argument("--img-loader", type=str, default='default', help="Image Loader")
     parser.add_argument("--epoch", type=int, default=100, help="Epoch number.")
     parser.add_argument("--batch-size", type=int, default=256, help="Batch size.")
-    parser.add_argument("--lr", type=int, default=1e-4, help="Learning rate.")
+    parser.add_argument("--lr", type=float, default=1e-4, help="Learning rate.")
     parser.add_argument("--img-size", type=int, default=224,
                         help="Height Weight of the training images after transform.")
     parser.add_argument('--load-model', action="store_true", default=False)
